Domains: Fix pagination bug with featured domains
There is a bug where the featured suggestions could fail to render due to being indexed outside of the initial pagination (ex. Recommended suggestion’s index > `PAGE_SIZE`).

This change fixes this by moving both the recommended and best alternative suggestions to the first and second index, respectively.

It also refactors marking featured domains outside of the component and into a utility function.