Activity Log: Use `getHttpData()` to get log events
Activity Log events don't interact with other data in Calypso and
it's all read-only; it doesn't listen for any other actions other
than the one that initially injects it into state.

Because of this there is a lot of basic machinery in place which
supports that simple fetching. However, that also makes it a
perfect candidate for beta testing with the new `getHttpData()`
helper function which removes the "boilerplate"

In this patch I'm removing the Redux and traditional data-layer
machinery in favor of the `requestHttpData()` getter and system.